[CrmSettingsFactory] Refactoring of the DefaultServiceWithSettings to allow classes other than IServices to retrieve CrmSettings
